@@ -7,14 +7,14 @@ if(versionPropsFile.exists())
 def code = (versionProps['VERSION_CODE'] ?: "1").toInteger()
 
 android {
-    compileSdkVersion 31
+    compileSdkVersion 33
     defaultConfig {
         applicationId "net.yasmar.crond"
         minSdkVersion 26
-        targetSdkVersion 31
+        targetSdkVersion 33
         versionCode code
         versionName "v1.3 ("+code+")"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
     }
     namespace 'net.yasmar.crond'
     buildTypes {
@@ -28,12 +28,13 @@ android {
 repositories {
     google()
     mavenCentral()
-    jcenter()
+    maven { url 'https://jitpack.io' }
 }
 
 dependencies {
-    api 'com.android.support:appcompat-v7:28.0.0'
+    api 'androidx.appcompat:appcompat:1.5.1'
     api 'com.cronutils:cron-utils:4.1.3'
-    api 'eu.chainfire:libsuperuser:1.0.0.201704021214'
-    api 'com.google.guava:guava:22.0-android'
+    implementation 'eu.chainfire:libsuperuser:1.1.1'
+    // Google messed up something (duplicate class error), apparently this fixes it
+    implementation 'com.google.guava:listenablefuture:9999.0-empty-to-avoid-conflict-with-guava'
 }
