@@ -1,20 +1,21 @@
 package net.yasmar.crond;
 
+import android.Manifest;
+import android.app.NotificationChannel;
 import android.app.NotificationManager;
 import android.app.PendingIntent;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
+import android.content.pm.PackageManager;
 import android.net.Uri;
-import android.os.AsyncTask;
+import android.os.Build;
 import android.os.Bundle;
 import android.os.Handler;
+import android.os.Looper;
 import android.os.PowerManager;
 import android.provider.Settings;
-import android.support.v4.app.NotificationCompat;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.app.AppCompatActivity;
 import android.text.Layout;
 import android.text.method.ScrollingMovementMethod;
 import android.view.View;
@@ -23,11 +24,16 @@ import android.widget.CheckBox;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 
-import java.lang.ref.WeakReference;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 
+import androidx.appcompat.app.AlertDialog;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.app.NotificationCompat;
 import eu.chainfire.libsuperuser.Shell;
 
 import static android.view.View.VISIBLE;
+import static net.yasmar.crond.Constants.POST_NOTIFICATIONS;
 import static net.yasmar.crond.Constants.PREFERENCES_FILE;
 import static net.yasmar.crond.Constants.PREF_BATTERY_WARNING;
 import static net.yasmar.crond.Constants.PREF_ENABLED;
@@ -39,7 +45,10 @@ public class MainActivity extends AppCompatActivity {
 
     private static final String TAG = "MainActivity";
 
-    private final Handler refreshHandler = new Handler();
+    private final ExecutorService executor = Executors.newSingleThreadExecutor();
+    private final Handler handler = new Handler(Looper.getMainLooper());
+
+    private final Handler refreshHandler = new Handler(Looper.getMainLooper());
 
     private Crond crond = null;
 
@@ -50,55 +59,44 @@ public class MainActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-        new RootChecker(this).execute();
+        sharedPrefs = getSharedPreferences(PREFERENCES_FILE, MODE_PRIVATE);
+        checkRoot();
     }
 
-    private static class RootChecker extends AsyncTask<Void, Void, Boolean> {
-        private WeakReference<MainActivity> contextRef = null;
-
-        public RootChecker(MainActivity context) {
-            contextRef = new WeakReference<>(context);
-        }
-        @Override
-        protected Boolean doInBackground(Void... params) {
+    private void checkRoot() {
+        executor.execute(() -> {
             IO.rootAvailable = Shell.SU.available();
-            IO.nonRootPrefix = contextRef.get().getExternalFilesDir(null);
-            return IO.rootAvailable;
-        }
+            IO.nonRootPrefix = getExternalFilesDir(null);
+            handler.post(() -> {
+                postCheckRoot();
+            });
+        });
+    }
 
-        @Override
-        protected void onPostExecute(Boolean rootAvail) {
-            MainActivity context = contextRef.get();
-            if (!IO.rootAvailable) {
-                context.noRoot();
-            } else {
-                context.init();
+    private void postCheckRoot() {
+        final Runnable next = this::checkBattery;
+        if (!IO.rootAvailable) {
+            boolean hasWarned = sharedPrefs.getBoolean(PREF_ROOT_WARNING, false);
+            if (!hasWarned) {
+                new AlertDialog.Builder(this)
+                        .setTitle("Root not detected")
+                        .setMessage("Without root, functionality will be extremely limited.")
+                        .setNeutralButton(android.R.string.ok, (d, w) -> {
+                            SharedPreferences.Editor editor = sharedPrefs.edit();
+                            editor.putBoolean(PREF_ROOT_WARNING, true);
+                            editor.apply();
+                            next.run();
+                        })
+                        .setIcon(android.R.drawable.ic_dialog_alert)
+                        .show();
+                return;
             }
         }
-    }
-
-    private void noRoot() {
-        sharedPrefs = getSharedPreferences(PREFERENCES_FILE, MODE_PRIVATE);
-        boolean hasWarned = sharedPrefs.getBoolean(PREF_ROOT_WARNING, false);
-        if (!hasWarned) {
-            new AlertDialog.Builder(this)
-                    .setTitle("Root not detected")
-                    .setMessage("Without root, functionality will be extremely limited. Note that crond should be fully closed after enabling root.")
-                    .setNeutralButton(android.R.string.ok, (d, w) -> {
-                        SharedPreferences.Editor editor = sharedPrefs.edit();
-                        editor.putBoolean(PREF_ROOT_WARNING, true);
-                        editor.apply();
-                        checkBattery();
-                    })
-                    .setIcon(android.R.drawable.ic_dialog_alert)
-                    .show();
-        } else {
-            checkBattery();
-        }
+        next.run();
     }
 
     private void checkBattery() {
-        sharedPrefs = getSharedPreferences(PREFERENCES_FILE, MODE_PRIVATE);
+        final Runnable next = this::checkNotifications;
         boolean hasWarned = sharedPrefs.getBoolean(PREF_BATTERY_WARNING, false);
         String packageName = getPackageName();
         PowerManager pm = (PowerManager)getSystemService(POWER_SERVICE);
@@ -114,17 +112,31 @@ public class MainActivity extends AppCompatActivity {
                         intent.setAction(Settings.ACTION_REQUEST_IGNORE_BATTERY_OPTIMIZATIONS);
                         intent.setData(Uri.parse("package:"+packageName));
                         startActivity(intent);
-                        init();
+                        next.run();
                     })
                     .setNegativeButton(android.R.string.no, (d, w) -> {
                         SharedPreferences.Editor editor = sharedPrefs.edit();
                         editor.putBoolean(PREF_BATTERY_WARNING, true);
                         editor.apply();
-                        init();
+                        next.run();
                     })
                     .setIcon(android.R.drawable.ic_dialog_alert)
                     .show();
-        } else {
+            return;
+        }
+        next.run();
+    }
+
+    private void checkNotifications() {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+            boolean showNotification = sharedPrefs.getBoolean(PREF_NOTIFICATION_ENABLED, false);
+            if (showNotification) {
+                if (checkSelfPermission(Manifest.permission.POST_NOTIFICATIONS) != PackageManager.PERMISSION_GRANTED) {
+                    requestPermissions(new String[]{Manifest.permission.POST_NOTIFICATIONS}, POST_NOTIFICATIONS);
+                }
+            }
+        }
+        if (!inited) {
             init();
         }
     }
@@ -132,7 +144,6 @@ public class MainActivity extends AppCompatActivity {
     private void init() {
         inited = true;
         crond = new Crond(this);
-        sharedPrefs = getSharedPreferences(PREFERENCES_FILE, MODE_PRIVATE);
 
         LinearLayout layout = findViewById(R.id.root_layout);
         for (int i = 0; i<layout.getChildCount(); i++) {
@@ -155,55 +166,44 @@ public class MainActivity extends AppCompatActivity {
         final CheckBox notificationCheckBox = findViewById(
                 R.id.check_notification_setting);
         notificationCheckBox.setChecked(sharedPrefs.getBoolean(PREF_NOTIFICATION_ENABLED, false));
-        notificationCheckBox.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                sharedPrefs.edit().putBoolean(PREF_NOTIFICATION_ENABLED,
-                        notificationCheckBox.isChecked()).apply();
+        notificationCheckBox.setOnClickListener(view -> {
+            sharedPrefs.edit().putBoolean(PREF_NOTIFICATION_ENABLED,
+                    notificationCheckBox.isChecked()).apply();
+            if (notificationCheckBox.isChecked()) {
+                checkNotifications();
             }
         });
 
         final CheckBox wakeLockCheckBox = findViewById(
                 R.id.check_wakelock_setting);
         wakeLockCheckBox.setChecked(sharedPrefs.getBoolean(PREF_USE_WAKE_LOCK, false));
-        wakeLockCheckBox.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                sharedPrefs.edit().putBoolean(PREF_USE_WAKE_LOCK,
-                        wakeLockCheckBox.isChecked()).apply();
-            }
-        });
+        wakeLockCheckBox.setOnClickListener(v -> sharedPrefs.edit().putBoolean(PREF_USE_WAKE_LOCK,
+                wakeLockCheckBox.isChecked()).apply());
 
         final Button enableButton = findViewById(R.id.button_enable);
-        enableButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                boolean oldEnabled = sharedPrefs.getBoolean(PREF_ENABLED, false);
-                sharedPrefs.edit().putBoolean(PREF_ENABLED, !oldEnabled).apply();
-                updateEnabled();
-                crond.scheduleCrontab(true, false, false);
-                refreshImmediately();
-            }
+        enableButton.setOnClickListener(view -> {
+            boolean oldEnabled = sharedPrefs.getBoolean(PREF_ENABLED, false);
+            sharedPrefs.edit().putBoolean(PREF_ENABLED, !oldEnabled).apply();
+            updateEnabled();
+            crond.scheduleCrontab(true, false, false);
+            refreshImmediately();
         });
 
         final Button clearButton = findViewById(R.id.button_clear_log);
-        clearButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                new AlertDialog.Builder(MainActivity.this)
-                        .setTitle(R.string.dialog_clean_title)
-                        .setMessage(R.string.dialog_clean_message)
-                        .setNegativeButton(R.string.no, null)
-                        .setPositiveButton(R.string.yes, new AlertDialog.OnClickListener() {
-                            @Override
-                            public void onClick(DialogInterface dialog, int which) {
-                                new LogClearer().execute();
-                                refreshImmediately();
-                            }
-                        })
-                        .show();
-            }
-        });
+        clearButton.setOnClickListener(view -> new AlertDialog.Builder(MainActivity.this)
+                .setTitle(R.string.dialog_clean_title)
+                .setMessage(R.string.dialog_clean_message)
+                .setNegativeButton(R.string.no, null)
+                .setPositiveButton(R.string.yes, new AlertDialog.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int which) {
+                        executor.execute(() -> {
+                            IO.clearLogFile();
+                        });
+                        refreshImmediately();
+                    }
+                })
+                .show());
 
         updateEnabled();
         refreshHandler.post(refresh);
@@ -211,11 +211,32 @@ public class MainActivity extends AppCompatActivity {
 
     @Override
     public void onResume() {
-       super.onResume();
+        super.onResume();
+        if (inited && !IO.rootAvailable) {
+            checkRootGranted();
+        }
         refreshHandler.removeCallbacksAndMessages(null);
         if (inited) {
             refreshHandler.post(refresh);
         }
+    }
+
+    private void checkRootGranted() {
+        executor.execute(() -> {
+            if (Shell.SU.available()) {
+                handler.post(() -> {
+                    new AlertDialog.Builder(this)
+                            .setTitle("Root granted")
+                            .setMessage("The app will close. Please re-open the app again.")
+                            .setNeutralButton(android.R.string.ok, (d, w) -> {
+                                finish();
+                                System.exit(0);
+                            })
+                            .setIcon(android.R.drawable.ic_dialog_alert)
+                            .show();
+                });
+            }
+        });
     }
 
     @Override
@@ -254,8 +275,11 @@ public class MainActivity extends AppCompatActivity {
         int flags = PendingIntent.FLAG_UPDATE_CURRENT | PendingIntent.FLAG_IMMUTABLE;
         PendingIntent pendingIntent = PendingIntent.getActivity(context, 0, intent,
                 flags);
+        NotificationChannel channel = new NotificationChannel("post_run_notification", "Post-run notification", NotificationManager.IMPORTANCE_LOW);
+        channel.setDescription("This channel is used for the post-run notification");
+        notificationManager.createNotificationChannel(channel);
         notificationManager.notify(1,
-                new NotificationCompat.Builder(context)
+                new NotificationCompat.Builder(context, "post_run_notification")
                         .setSmallIcon(R.drawable.ic_notification_icon)
                         .setColor(context.getColor(R.color.colorPrimary))
                         .setContentTitle(context.getString(R.string.app_name))
@@ -269,63 +293,27 @@ public class MainActivity extends AppCompatActivity {
     private final Runnable refresh = new Runnable() {
         @Override
         public void run() {
-            new FileReader(MainActivity.this).execute();
+            executor.execute(() -> {
+                crond.setCrontab(IO.readFileContents(IO.getCrontabPath()));
+                CharSequence crontab = crond.processCrontab();
+                CharSequence log = IO.readFileContents(IO.getLogPath());
+                handler.post(() -> {
+                    final TextView crontabContent = findViewById(R.id.text_content_crontab);
+                    crontabContent.setText(crontab);
+
+                    final TextView crondLog = findViewById(R.id.text_content_crond_log);
+                    crondLog.setText(log);
+                    Layout layout = crondLog.getLayout();
+                    if (layout == null) {
+                        // When launching with the screen off... we get here. Avoid crashing!
+                        return;
+                    }
+                    final int scrollAmount = layout.getLineTop(crondLog.getLineCount()) - crondLog.getHeight();
+                    // if there is no need to scroll, scrollAmount will be <=0
+                    crondLog.scrollTo(0, Math.max(scrollAmount, 0));
+                });
+            });
             refreshHandler.postDelayed(refresh, 10000);
         }
     };
-
-    private static class FileReader extends AsyncTask<Void, Void, CharSequence[]> {
-        private WeakReference<MainActivity> contextRef = null;
-
-        public FileReader(MainActivity context) {
-            contextRef = new WeakReference<>(context);
-        }
-
-        @Override
-        protected CharSequence[] doInBackground(Void... params) {
-            MainActivity context = contextRef.get();
-            if (context == null || context.isFinishing()) {
-                cancel(true);
-            }
-            CharSequence[] ret = new CharSequence[2];
-            Crond crond = context.crond;
-            crond.setCrontab(IO.readFileContents(IO.getCrontabPath()));
-            ret[0] = crond.processCrontab();
-
-            ret[1] = IO.readFileContents(IO.getLogPath());
-            return ret;
-        }
-
-        @Override
-        protected void onPostExecute(CharSequence[] sequences) {
-            MainActivity context = contextRef.get();
-            if (context == null || context.isFinishing()) {
-                return;
-            }
-            final TextView crontabContent = context.findViewById(R.id.text_content_crontab);
-            crontabContent.setText(sequences[0]);
-
-            final TextView crondLog = context.findViewById(R.id.text_content_crond_log);
-            crondLog.setText(sequences[1]);
-            Layout layout = crondLog.getLayout();
-            if (layout == null) {
-                // When launching with the screen off... we get here. Avoid crashing!
-                return;
-            }
-            final int scrollAmount = layout.getLineTop(crondLog.getLineCount()) - crondLog.getHeight();
-            // if there is no need to scroll, scrollAmount will be <=0
-            if (scrollAmount > 0)
-                crondLog.scrollTo(0, scrollAmount);
-            else
-                crondLog.scrollTo(0, 0);
-        }
-    }
-
-    private static class LogClearer extends AsyncTask<Void, Void, Void> {
-        @Override
-        protected Void doInBackground(Void... params) {
-            IO.clearLogFile();
-            return null;
-        }
-    }
 }
